Added email to admin, added method changeAdmin

+++ src/main/java/com/alex/courses/repository/AdminRepository.java
@@ -3,5 +3,5 @@ package com.alex.courses.repository;
 import com.alex.courses.entity.Administrator;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdminRepository extends JpaRepository<Administrator, Integer> {
+public interface AdminRepository extends JpaRepository<Administrator, Long> {
 }
